# wsfl-frontend/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "WSFL (Winter Series Fantasy League) API",
        "description": "API for managing fantasy running league competitions",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://choosealicense.com/licenses/mit/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/api/teams/{id}": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Get team by ID",
                "description": "Retrieves detailed information about a specific team",
                "operationId": "getTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the team to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Team not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Team not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Team found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamDetailsDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Team"
                ],
                "summary": "Update team",
                "description": "Updates an existing team's information",
                "operationId": "updateTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the team to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamUpdateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Team not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Team not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Team updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamDetailsDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to update teams",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Team"
                ],
                "summary": "Delete team",
                "description": "Deletes an existing team and its roster",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the team to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Team not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Team not found with id: 123"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to delete teams",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Team deleted successfully"
                    }
                }
            }
        },
...
        "/api/teams": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Get all teams",
                "description": "Retrieves a paginated list of all teams in the system",
                "operationId": "getAllTeams",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved teams",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageTeamSummaryDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Create new team",
                "description": "Creates a new team in a specific league",
                "operationId": "createTeam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Team created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamDetailsDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid input data"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "League not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "League not found with id: 1"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to create teams",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teams/{id}/runners": {
            "get": {
                "tags": [
                    "Team"
                ],
                "summary": "Get team runners",
                "description": "Retrieves all runners on a specific team's roster",
                "operationId": "getTeamRunners",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Team not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Team not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully retrieved runners",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RunnerSummaryDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TeamSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Team's unique identifier",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Team's name",
                        "example": "Thunderbolts"
                    },
                    "wins": {
                        "type": "integer",
                        "description": "Number of wins",
                        "format": "int32",
                        "example": 10
                    },
                    "losses": {
                        "type": "integer",
                        "description": "Number of losses",
                        "format": "int32",
                        "example": 5
                    },
                    "ties": {
                        "type": "integer",
                        "description": "Number of ties",
                        "format": "int32",
                        "example": 2
                    },
                    "totalScore": {
                        "type": "integer",
                        "description": "Total score",
                        "format": "int32",
                        "example": 157
                    }
                },
                "description": "Summary representation of a team"
            },
...
            "TeamUpdateDto": {
                "required": [
                    "city",
                    "leagueId",
                    "name",
                    "state"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 2,
                        "type": "string",
                        "description": "The name of the team",
                        "example": "Swift Striders"
                    },
                    "leagueId": {
                        "type": "integer",
                        "description": "The ID of the league this team belongs to",
                        "format": "int64",
                        "example": 1
                    },
                    "city": {
                        "maxLength": 100,
                        "minLength": 2,
                        "type": "string",
                        "description": "The team's home city",
                        "example": "Boston"
                    },
                    "state": {
                        "maxLength": 2,
                        "minLength": 2,
                        "type": "string",
                        "description": "The team's home state",
                        "example": "MA"
                    }
                },
                "description": "Updated team details"
            },
            "LeagueSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the league",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the league"
                    },
                    "season": {
                        "type": "integer",
                        "description": "Current season number",
                        "format": "int32"
                    }
                },
                "description": "Summary information about a league"
            },
            "RunnerSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Runner's unique identifier",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Runner's full name",
                        "example": "John Doe"
                    },
                    "gender": {
                        "type": "string",
                        "description": "Runner's gender (M or F)",
                        "example": "M"
                    }
                },
                "description": "Summary representation of a runner"
            },
            "TeamDetailsDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the team",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the team",
                        "example": "Speed Demons"
                    },
                    "wins": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Number of matches won",
                        "format": "int32",
                        "example": 5
                    },
                    "losses": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Number of matches lost",
                        "format": "int32",
                        "example": 3
                    },
                    "ties": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Number of matches tied",
                        "format": "int32",
                        "example": 1
                    },
                    "totalScore": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Total score accumulated by the team",
                        "format": "int32",
                        "example": 450
                    },
                    "league": {
                        "$ref": "#/components/schemas/LeagueSummaryDto"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/UserSummaryDto"
                    },
                    "runners": {
                        "type": "array",
                        "description": "List of runners on the team",
                        "items": {
                            "$ref": "#/components/schemas/RunnerSummaryDto"
                        }
                    }
                },
                "description": "Detailed representation of a team"
            },
            "UserSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the user",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Full name of the user",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email address of the user",
                        "example": "john.doe@example.com"
                    }
                },
                "description": "Summary representation of a user"
            },
...
            "TeamCreateDto": {
                "required": [
                    "leagueId",
                    "name",
                    "ownerId"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string",
                        "description": "Name of the team",
                        "example": "Speed Demons"
                    },
                    "leagueId": {
                        "type": "integer",
                        "description": "ID of the league this team belongs to",
                        "format": "int64",
                        "example": 1
                    },
                    "ownerId": {
                        "type": "integer",
                        "description": "ID of the team owner/manager",
                        "format": "int64",
                        "example": 1
                    }
                },
                "description": "Team creation details"
            },
...
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            },
            "PageTeamSummaryDto": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeamSummaryDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}